refactor(sellers): unify discover + /all on shared featuredServiceColumns
Promote the polished discover Featured Services columns into a shared
module under sellers/featured-columns.tsx and have both consumers
import it:

- /all's Top Servers (KnownSellersTable)
- /'s Featured Services (DiscoverSellersTable)
- /'s search results table (DiscoverSearchResults)

Effect on /all: gains the title + description + hostname server cell
stack, the wider activity sparkline, the promoted Volume column, and
the "Try it" action button. Loses the Facilitator column (was already
duplicated on /facilitators) and the old single-line Origins/Address
server cell.

Drops the now-orphaned known-sellers/columns.tsx and the unreferenced
OriginsSkeleton export from origins.tsx (its only consumer was the
old known-sellers columns module).
