+++ apps/scan/src/app/(app)/(home)/(discover)/_components/discover-origins.tsx
@@ -6,7 +6,10 @@ import { useSellersSorting } from '@/app/(app)/_contexts/sorting/sellers/hook';
 import { useTimeRangeContext } from '@/app/(app)/_contexts/time-range/hook';
 import { useChain } from '@/app/(app)/_contexts/chain/hook';
 
-import { discoverColumns as columns, type DiscoverColumnType } from './columns';
+import {
+  featuredServiceColumns as columns,
+  type FeaturedServiceItem,
+} from '@/app/(app)/(home)/(overview)/_components/sellers/featured-columns';
 import { api } from '@/trpc/client';
 
 interface Props {
@@ -31,7 +34,7 @@ export const DiscoverSellersTable: React.FC<Props> = ({ originUrls }) => {
   return (
     <DataTable
       columns={columns}
-      data={topSellers.items as DiscoverColumnType[]}
+      data={topSellers.items as FeaturedServiceItem[]}
       pageSize={15}
     />
   );
